Deploy_api_and_Monitoring/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
pydantic==2.11.5
prometheus-fastapi-instrumentator==7.1.0
prometheus-client==0.22.1
python-multipart==0.0.9
